Open form on start container
