+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2022-07-07T07:42:01.312Z\n"
-"PO-Revision-Date: 2022-07-07T07:42:01.312Z\n"
+"POT-Creation-Date: 2022-07-07T08:35:40.367Z\n"
+"PO-Revision-Date: 2022-07-07T08:35:40.367Z\n"
 
 msgid "API error code: {{statusText}} ({{status}})"
 msgstr ""
@@ -32,7 +32,7 @@ msgstr ""
 msgid "List"
 msgstr ""
 
-msgid "Image created successfully"
+msgid "Image started successfully"
 msgstr ""
 
 msgid "Create new container"
@@ -45,24 +45,6 @@ msgid "Submit"
 msgstr ""
 
 msgid "Cancel"
-msgstr ""
-
-msgid "Search by name"
-msgstr ""
-
-msgid "Name"
-msgstr ""
-
-msgid "Status"
-msgstr ""
-
-msgid "Open DHIS2 instance"
-msgstr ""
-
-msgid "Starting container(s)"
-msgstr ""
-
-msgid "Image(s) started successfully"
 msgstr ""
 
 msgid "Stopping container(s)"
@@ -81,6 +63,18 @@ msgid "Stop container"
 msgstr ""
 
 msgid "Details"
+msgstr ""
+
+msgid "Search by name"
+msgstr ""
+
+msgid "Name"
+msgstr ""
+
+msgid "Open DHIS2 instance"
+msgstr ""
+
+msgid "Status"
 msgstr ""
 
 msgid "Drag and drop some files here, or click to select files"
