+++ i18n/es.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2022-07-07T07:42:01.312Z\n"
+"POT-Creation-Date: 2022-07-07T08:35:40.367Z\n"
 "PO-Revision-Date: 2018-10-25T09:02:35.143Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -32,7 +32,7 @@ msgstr "Añadir"
 msgid "List"
 msgstr "Listar"
 
-msgid "Image created successfully"
+msgid "Image started successfully"
 msgstr ""
 
 msgid "Create new container"
@@ -45,24 +45,6 @@ msgid "Submit"
 msgstr ""
 
 msgid "Cancel"
-msgstr ""
-
-msgid "Search by name"
-msgstr ""
-
-msgid "Name"
-msgstr ""
-
-msgid "Status"
-msgstr ""
-
-msgid "Open DHIS2 instance"
-msgstr ""
-
-msgid "Starting container(s)"
-msgstr ""
-
-msgid "Image(s) started successfully"
 msgstr ""
 
 msgid "Stopping container(s)"
@@ -81,6 +63,18 @@ msgid "Stop container"
 msgstr ""
 
 msgid "Details"
+msgstr ""
+
+msgid "Search by name"
+msgstr ""
+
+msgid "Name"
+msgstr ""
+
+msgid "Open DHIS2 instance"
+msgstr ""
+
+msgid "Status"
 msgstr ""
 
 msgid "Drag and drop some files here, or click to select files"
